app/requirements.txt
asgiref==3.7.2
Django==4.2.1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
psycopg2==2.9.9
Pillow==10.1.0
djangorestframework==3.14.0
redis==5.0.1